AmsApi/requirements.txt
asgiref==3.6.0
Babel==2.12.1
backports.zoneinfo==0.2.1
black==23.3.0
click==8.1.3
Django==4.2
django-cors-headers==3.14.0
django-environ==0.10.0
django-money==3.0.0
djangorestframework==3.14.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
Pillow==10.0.0
platformdirs==3.2.0
psycopg2==2.9.6
py-moneyed==2.0
pypng==0.20220715.0
pytz==2023.3
qrcode==7.4.2
sqlparse==0.4.4
tomli==2.0.1
typing-extensions==4.5.0
